% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedhash.R
\name{create_seedhash}
\alias{create_seedhash}
\title{Create a SeedHash Generator (Convenience Function)}
\usage{
create_seedhash(input_string, min_value = 0, max_value = 2^31 - 1)
}
\arguments{
\item{input_string}{The string to hash for seed generation}

\item{min_value}{Minimum value for random number range (default: 0)}

\item{max_value}{Maximum value for random number range (default: 2^31 - 1)}
}
\value{
A new SeedHashGenerator object
}
\description{
This is a convenience function to create a SeedHashGenerator object.
}
\examples{
\dontrun{
gen <- create_seedhash("Shalini")
seeds <- gen$generate_seeds(10)
}

}
