% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedhash.R
\name{SeedHashGenerator}
\alias{SeedHashGenerator}
\title{SeedHash Generator}
\description{
SeedHash Generator

SeedHash Generator
}
\details{
Generate deterministic random seeds from string input using MD5 hashing.
This class allows users to generate reproducible sequences of random numbers
by hashing an input string and using it as a seed for R's random number generator.
}
\examples{
\donttest{
# Create a generator with default range
gen <- SeedHashGenerator$new("Shalini")

# Generate 10 random seeds
seeds <- gen$generate_seeds(10)
print(seeds)

# Create a generator with custom range
gen_custom <- SeedHashGenerator$new("MyProject", min_value = 1, max_value = 100)
seeds_custom <- gen_custom$generate_seeds(5)
print(seeds_custom)

# Get the MD5 hash
hash_value <- gen$get_hash()
print(hash_value)
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeedHashGenerator-new}{\code{SeedHashGenerator$new()}}
\item \href{#method-SeedHashGenerator-generate_seeds}{\code{SeedHashGenerator$generate_seeds()}}
\item \href{#method-SeedHashGenerator-get_hash}{\code{SeedHashGenerator$get_hash()}}
\item \href{#method-SeedHashGenerator-print}{\code{SeedHashGenerator$print()}}
\item \href{#method-SeedHashGenerator-clone}{\code{SeedHashGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeedHashGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeedHashGenerator-new}{}}}
\subsection{Method \code{new()}}{
The input string for seed generation


Minimum value for random number range


Maximum value for random number range


The integer seed derived from hashing


Initialize the SeedHashGenerator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeedHashGenerator$new(input_string, min_value = 0, max_value = 2^31 - 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_string}}{The string to hash for seed generation}

\item{\code{min_value}}{Minimum value for random number range (default: 0)}

\item{\code{max_value}}{Maximum value for random number range (default: 2^31 - 1)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new SeedHashGenerator object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeedHashGenerator-generate_seeds"></a>}}
\if{latex}{\out{\hypertarget{method-SeedHashGenerator-generate_seeds}{}}}
\subsection{Method \code{generate_seeds()}}{
Generate a list of random seed numbers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeedHashGenerator$generate_seeds(count)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{count}}{The number of random seeds to generate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of random integers within the specified range
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeedHashGenerator-get_hash"></a>}}
\if{latex}{\out{\hypertarget{method-SeedHashGenerator-get_hash}{}}}
\subsection{Method \code{get_hash()}}{
Get the MD5 hash of the input string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeedHashGenerator$get_hash()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The MD5 hash as a hexadecimal string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeedHashGenerator-print"></a>}}
\if{latex}{\out{\hypertarget{method-SeedHashGenerator-print}{}}}
\subsection{Method \code{print()}}{
Print method for SeedHashGenerator
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeedHashGenerator$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments (unused)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeedHashGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeedHashGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeedHashGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
